feat: Information section in Recruit page
